requirements.txt
numpy == 1.22.4
pillow == 9.1.1
pyyaml == 6.0
rstr == 3.2.0